+++ packages/core/src/common/tenant/tenant-protocol.ts
@@ -1,5 +1,0 @@
-export const TenantProvider = Symbol('TenantProvider');
-
-export interface TenantProvider {
-    provide(tenant?: string): Promise<string>;
-}